Take Some Parameters by Const Reference
Cool project-great work!
Sorry about the whitespace changes.

Change 1: Players_turn(). Since you aren't changing the value of note, you can take a const reference to it. The way it was before, the content of note was getting copied to memory, which means it's duplicated because it already exists in memory. This is a small optimization here, but is very important in game programming where memory allocation means a lot.

Change 2: If_win. I did the same thing here. Vectors can take up considerale amounts of space in memory. You already have the vector when you call If_win, but since you aren't taking it by const reference C++ will create ANOTHER copy of it in memory.

Change 3: Print_field. Same thing here

Change4:  I used the auto keyword for string construction. It's more of a styling, but it's considered to be the modern way.